Create Model 3
In this model, we minimize the costs due to flight duration, ground holding costs, speed deviations, level alterations, and cancellations. All the flights will be able to depart during a time frame  
 
from their scheduled departure time until the maximum departure waiting time otherwise the flight will be canceled. Also, the flight will be canceled if the cancellation cost of a flight is less than the estimated costs of the flight realization in the extreme scenario of heavy air congestion where the flight will have to exceed the maximum flight duration.